change: Patch vanilla chunk manager instead of rolling our own
The custom hash table based implementation was not sound. This might also
fix a few issues with mod compatibility. Should still be faster than the
vanilla implementation, and it will significantly reduce the amount of code
we're maintaining.
