+++ src/main/resources/sodium.mixins.json
@@ -27,6 +27,8 @@
     "features.buffer_builder.intrinsics.MixinSpriteTexturedVertexConsumer",
     "features.buffer_builder.intrinsics.MixinWorldRenderer",
     "features.chunk_rendering.MixinChunkBuilder",
+    "features.chunk_rendering.MixinClientChunkManager",
+    "features.chunk_rendering.MixinClientChunkManager$MixinClientChunkMap",
     "features.chunk_rendering.MixinClientWorld",
     "features.chunk_rendering.MixinWorldRenderer",
     "features.debug.MixinDebugHud",
